Implemented custom data type

+++ hw04/src/decimals.h
@@ -2,6 +2,6 @@
 #define DECIMALS_H_
 
 int decimals_to_base(int decimals);
-int load_decimal(const char *string, int decimals);
+long load_decimal(const char *string, int decimals);
 
 #endif
